move store locator contants
